Refactor: move `ForkChanges` to independent module
